Added compute leaderboard to admin api
